Logic have been changed to shift responsibility for operations on object to this objects.

+++ PerfumeStore.API/PerfumeStore.Core/Services/ProductsService.cs
@@ -1,7 +1,7 @@
 ﻿using PerfumeStore.Core.CustomExceptions;
 using PerfumeStore.Core.Repositories;
-using PerfumeStore.Core.RequestForms;
 using PerfumeStore.Domain.DbModels;
+using PerfumeStore.Domain.Models;
 
 namespace PerfumeStore.Core.Services
 {
@@ -18,19 +18,30 @@ namespace PerfumeStore.Core.Services
 
         public async Task<Product> CreateProductAsync(CreateProductForm createProductForm)
         {
-            ICollection<ProductCategory> productCategories = await _productCategoriesRepository.GetByIdsAsync(createProductForm.ProductCategoryId);
-            if (!productCategories.Any())
+            ICollection<ProductCategory> productCategories = await _productCategoriesRepository.GetByIdsAsync(createProductForm.ProductCategoriesIds);
+            if (productCategories.Count != createProductForm.ProductCategoriesIds.Count)
             {
-                throw new EntityNotFoundException<ProductCategory, int>("Product categories are missing.");
-            }
-            Product newProduct = GenerateNewProduct(createProductForm, productCategories);
-            newProduct = await _productsRepository.CreateAsync(newProduct);
+                var foundIds = productCategories.Select(pc => pc.Id);
+                var notFoundIds = createProductForm.ProductCategoriesIds.Except(foundIds);
+                string notFoundIdsString = string.Join(", ", notFoundIds);
 
-            return newProduct;
+                throw new EntityNotFoundException<ProductCategory, int>($"Can't find entities {typeof(ProductCategory)} with Ids: {notFoundIdsString}");
+            }
+
+            Product product = new Product();
+            product.CreateProduct(createProductForm, productCategories);
+            product = await _productsRepository.CreateAsync(product);
+
+            return product;
         }
 
         public async Task DeleteProductAsync(int productId)
         {
+            Product? product = await _productsRepository.GetByIdAsync(productId);
+            if (product == null)
+            {
+                throw new EntityNotFoundException<Product, int>($"Entity of type: {typeof(Product)} is missing. Entity Id: {productId}");
+            }
             await _productsRepository.DeleteAsync(productId);
         }
 
@@ -42,7 +53,6 @@ namespace PerfumeStore.Core.Services
 
         public async Task<Product> GetProductByIdAsync(int productId)
         {
-
             Product? product = await _productsRepository.GetByIdAsync(productId);
             if (product is null)
             {
@@ -53,38 +63,28 @@ namespace PerfumeStore.Core.Services
 
         public async Task<Product> UpdateProductAsync(UpdateProductForm updateForm)
         {
-            Product? productToUpdate = await _productsRepository.GetByIdAsync(updateForm.ProductId);
-            if (productToUpdate is null)
+            Product? product = await _productsRepository.GetByIdAsync(updateForm.ProductId);
+            if (product is null)
             {
                 throw new EntityNotFoundException<Product, int>($"Can't find entity {typeof(Product)} with Id: {updateForm.ProductId}");
             }
-            Product updatedProductId = await _productsRepository.UpdateAsync(productToUpdate);
 
-            return updatedProductId;
-        }
-
-        private Product GenerateNewProduct(CreateProductForm createProductForm, ICollection<ProductCategory> productCategories)
-        {
-            var productToCreate = new Product
+            ICollection<ProductCategory> newProductCategories = await _productCategoriesRepository.GetByIdsAsync(updateForm.ProductCategoriesIds);
+            if (newProductCategories.Count != updateForm.ProductCategoriesIds.Count)
             {
-                Name = createProductForm.ProductName,
-                Price = createProductForm.ProductPrice,
-                Description = createProductForm.ProductDescription,
-                ProductCategories = productCategories,
-                Manufacturer = createProductForm.ProductManufacturer,
-                DateAdded = DateTime.Now
-            };
+                var foundIds = newProductCategories.Select(pc => pc.Id);
+                var notFoundIds = updateForm.ProductCategoriesIds.Except(foundIds);
+                string notFoundIdsString = string.Join(", ", notFoundIds);
 
-            return productToCreate;
+                throw new EntityNotFoundException<ProductCategory, int>($"Can't find entities {typeof(ProductCategory)} with Ids: {notFoundIdsString}");
+            }
+
+            product.UpdateProduct(updateForm, newProductCategories);
+            product = await _productsRepository.UpdateAsync(product);
+
+            return product;
         }
 
-        private static void GenerateUpdatedProduct(UpdateProductForm updateform, Product productToUpdate)
-        {
-            productToUpdate.Name = updateform.ProductName;
-            productToUpdate.Price = updateform.ProductPrice;
-            productToUpdate.Description = updateform.ProductDescription;
-            productToUpdate.Manufacturer = updateform.ProductManufacturer;
-        }
     }
 }
